+++ sbm/oem/landrover.py
@@ -2,6 +2,8 @@
 Land Rover OEM handler for the SBM tool.
 """
 
+from typing import Any
+
 from .base import BaseOEMHandler
 
 
@@ -47,4 +49,34 @@ class LandRoverHandler(BaseOEMHandler):
             r"land-rover",
             r"land_rover",
             r"land rover",
+        ]
+
+    def get_predetermined_inside_style_configs(self) -> list[dict[str, Any]]:
+        """
+        Land Rover-specific CommonTheme styles to append to sb-inside.scss
+        when source SCSS indicates they were used.
+        """
+        return [
+            {
+                "label": "Land Rover Inside Pages Styles",
+                "source_indicator_file": "inside.scss",
+                "indicator_patterns": [
+                    r"_inside-pages-v2\.scss",
+                    r"_inside-pages\.scss",
+                ],
+                "common_theme_candidates": [
+                    "css/dealer-groups/landrover/_inside-pages-v2.scss",
+                    "css/dealer-groups/landrover/_inside-pages.scss",
+                ],
+                "dedupe_markers": ["#cpo-offers", ".offer-disclaimer", ".offers {"],
+            },
+            {
+                "label": "Land Rover National Offers Styles",
+                "source_indicator_file": "style.scss",
+                "indicator_patterns": [r"_national-offers\.scss"],
+                "common_theme_candidates": [
+                    "css/dealer-inspire-plugins/_national-offers.scss",
+                ],
+                "dedupe_markers": [".national-incentive-offers"],
+            },
         ]